Add files via upload
This is a compressed CSV file containing the IDs, grant years and CPC classifications of a million randomly chosen US patents granted between 1976 and 2022, It is used as a benchmark for testing automated patent classification methods. Due to the large size of patent texts,they are not included here but can be retrieved in bulk from www.patentsview.org. The training set for this benchmark is the set of all US patents between 1976 and 2022 which is also not included for its large size but can be easily retrieved from the aforementioned website. Note that the classification included in this dataset is for the lowest CPC level (subgroup) however the higher order classifications can be obtained from it using a simple text manipulation script.